% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rarity}
\alias{rarity}
\title{Run a RARity analysis.
\code{dir} is the directory where the data is stored.
\code{pheno_file} is a character vector of normalized phenotype file names, relative to \code{dir}.
for chromosome \code{i}.
Returns a data frame with the results.}
\usage{
rarity(dir, phenos)
}
\description{
Run a RARity analysis.
\code{dir} is the directory where the data is stored.
\code{pheno_file} is a character vector of normalized phenotype file names, relative to \code{dir}.
for chromosome \code{i}.
Returns a data frame with the results.
}
