% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{set_blocks_per_chunk}
\alias{set_blocks_per_chunk}
\title{Set the number of blocks per chunk.
\code{blocks_per_chunk} is the number of blocks per chunk.}
\usage{
set_blocks_per_chunk(blocks_per_chunk)
}
\description{
Set the number of blocks per chunk.
\code{blocks_per_chunk} is the number of blocks per chunk.
}
