% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{set_min_sum}
\alias{set_min_sum}
\title{Set the min sum for the RARity analysis.
\code{min_sum} is the min sum.}
\usage{
set_min_sum(min_sum)
}
\description{
Set the min sum for the RARity analysis.
\code{min_sum} is the min sum.
}
