% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{set_num_threads}
\alias{set_num_threads}
\title{Set the number of threads.
\code{num_threads} is the number of threads.}
\usage{
set_num_threads(num_threads)
}
\description{
Set the number of threads.
\code{num_threads} is the number of threads.
}
